+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* fixed refresh token when it expired
 * fixed retry at BAD_SESSION
 * added credentials authentication
 * added CI test